feat(ritk-filter): SpeckleNoise float-exact via ITK gamma over MT19937
Third RNG filter unblocked. Rewrote noise/speckle.rs to ITK's
multiplicative Gamma(mean 1, var std^2): Marsaglia-Tsang
acceptance-rejection for the fractional shape over the MT19937
GetVariateWithOpenUpperRange stream (int/2^32, added to noise/mersenne.rs),
plus floork exponential draws. Integer-shape (delta=0) handled by IEEE
1/0->inf exactly as ITK C++. seed u64->u32; std==0 guard (identity).

out = gamma * input. <=1e-3 vs sitk.SpeckleNoise single-threaded across
std in {0.3,0.5,0.7} (0.5 exercises delta=0). Coverage 271 -> 272/298.
Only ShotNoise (Poisson) remains of the noise family.

+++ crates/ritk-python/src/filter/noise.rs
@@ -116,8 +116,8 @@ pub fn shot_noise(py: Python<'_>, image: &PyImage, scale: f64, seed: u64) -> Rit
 /// Returns:
 ///     PyImage with speckle noise.
 #[pyfunction]
-#[pyo3(signature = (image, std, seed=42_u64))]
-pub fn speckle_noise(py: Python<'_>, image: &PyImage, std: f64, seed: u64) -> RitkResult<PyImage> {
+#[pyo3(signature = (image, std, seed=42_u32))]
+pub fn speckle_noise(py: Python<'_>, image: &PyImage, std: f64, seed: u32) -> RitkResult<PyImage> {
     let img = Arc::clone(&image.inner);
     let result = py
         .allow_threads(|| {
